feat(catalogo): add shared search param utilities
Extract parseParamArray and dedupeArray into utils/searchParams.ts.
Update filterUtils.ts, urlFilters.ts, and useCatalogFiltersStore.ts to
import and use the shared utilities for consistent normalization and
deduplication of filter values.
