+++ apps/web/src/app/(sections)/catalogo/utils/searchParams.ts
@@ -1,0 +1,9 @@
+import { normalizeString } from '@frijolmagico/utils/string'
+
+export const parseParamArray = (param?: string | null) =>
+  param && param.length > 0
+    ? param.split(',').filter(Boolean).map(normalizeString)
+    : []
+
+export const dedupeArray = (arr: string[]) =>
+  Array.from(new Set(arr.map(normalizeString)))
